requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
gspread>=5.12.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
plotly>=5.18.0
python-dateutil>=2.8.0
